Refactor reducer imports in locationAddEdit
Change also selectors to be reselect
