+++ packages/location-management/src/components/LocationUnitAddEdit/Form.tsx
@@ -20,7 +20,7 @@ import { Geometry } from 'geojson';
 import { ERROR_OCCURED, LOCATION_UNIT_POST_PUT } from '../../constants';
 import { v4 } from 'uuid';
 import { LocationUnitGroup } from '../../ducks/location-unit-groups';
-import { ParsedHierarchyNode } from '../../ducks/types';
+import { ParsedHierarchyNode, RawOpenSRPHierarchy } from '../../ducks/locationHierarchy/types';
 import { sendErrorNotification, sendSuccessNotification } from '@opensrp/notifications';
 import { Dictionary } from '@onaio/utils';
